+++ .gitignore
@@ -1,1 +1,3 @@
 node_modules/
+.env
+public/js/apiKeys.js
