--- config/.env
@@ -1,1 +1,0 @@
-JAWSDB_URL='mysql://xbuuk9gp3y04z8qc:t36vabtn99wbzb03@yhrz9vns005e0734.cbetxkdyhwsb.us-east-1.rds.amazonaws.com:3306/dj3lu0ivan1m31rr'
