+++ package.json
@@ -18,7 +18,9 @@
   },
   "homepage": "https://github.com/MarciaAdler/Project2#readme",
   "dependencies": {
+    "axios": "^0.19.2",
     "bcryptjs": "^2.4.3",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "express-handlebars": "^3.1.0",
     "express-session": "^1.17.0",
