+++ server.js
@@ -1,8 +1,14 @@
+require('dotenv').config();
 // Requiring necessary npm packages
 var express = require("express");
 var session = require("express-session");
+
+
+
 // Requiring passport as we've configured it
 var passport = require("./config/passport");
+
+
 
 // Setting up port and requiring models for syncing
 var PORT = process.env.PORT || 8080;
