+++ views/index.handlebars
@@ -1,9 +1,5 @@
 <div class="container-fluid">
-            <div class="row">
-                <div class="col-12 text-center">
-                    <div class="alert alert-primary">MAP?</div>
-                </div>
-            </div>
+            
             <div class="row">
                 <div class="col-12 col-lg-8">
                     <div class="line-graph">
